+++ tket2-hseries/src/extension/qsystem.rs
@@ -36,6 +36,7 @@ use strum::{EnumIter, EnumString, IntoStaticStr};
 
 use super::futures::future_type;
 
+mod barrier;
 mod lower;
 use lower::pi_mul_f64;
 pub use lower::{check_lowered, lower_tk2_op, LowerTk2Error, LowerTket2ToQSystemPass};
